chore: run all tests in all environments
Remove repo, block service and other unecessary deps.
