+++ package.json
@@ -65,12 +65,11 @@
     "npm": ">=3.0.0"
   },
   "dependencies": {
-    "cids": "multiformats/js-cid#fix/support-uint8arrays",
+    "cids": "^1.0.0",
     "class-is": "^1.1.0",
     "multicodec": "^2.0.0",
     "multihashing-async": "^2.0.0",
-    "npm": "^6.14.7",
-    "protons": "^1.2.1",
+    "protons": "^2.0.0",
     "reset": "^0.1.0",
     "run": "^1.4.0",
     "stable": "^0.1.8",
@@ -78,11 +77,6 @@
   },
   "devDependencies": {
     "aegir": "^25.0.0",
-    "fs-extra": "^9.0.1",
-    "ipfs-block-service": "^0.17.1",
-    "ipfs-repo": "^5.0.0",
-    "ipfs-utils": "^2.3.1",
-    "ipld-block": "^0.9.2",
     "multibase": "^3.0.0",
     "multihashes": "^3.0.0"
   }
