--- test/test-repo/blocks/SHARDING
@@ -1,1 +1,0 @@
-/repo/flatfs/shard/v1/next-to-last/2
