--- test/test-repo/datastore/CURRENT
@@ -1,1 +1,0 @@
-MANIFEST-000011
