--- test/test-repo/datastore/LOG
@@ -1,5 +1,0 @@
-=============== Sep 9, 2016 (EDT) ===============
-14:06:48.740952 log@legend F·NumFile S·FileSize N·Entry C·BadEntry B·BadBlock Ke·KeyError D·DroppedEntry L·Level Q·SeqNum T·TimeElapsed
-14:06:48.741276 db@open opening
-14:06:48.745559 db@janitor F·4 G·0
-14:06:48.745598 db@open done T·4.297502ms
