--- test/test-repo/datastore/LOG.old
@@ -1,5 +1,0 @@
-=============== Aug 17, 2016 (CEST) ===============
-16:48:13.278954 log@legend F·NumFile S·FileSize N·Entry C·BadEntry B·BadBlock Ke·KeyError D·DroppedEntry L·Level Q·SeqNum T·TimeElapsed
-16:48:13.280400 db@open opening
-16:48:13.285876 db@janitor F·4 G·0
-16:48:13.285922 db@open done T·5.49005ms
