--- test/test-repo/version
@@ -1,1 +1,0 @@
-6
